refactor structure of Calendar folder
